pyproject.toml
[project]
name = "pw-router"
version = "0.1.0"
description = "Minimal, auditable LLM routing gateway for regulated environments"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    {name = "Protocol Wealth LLC", email = "engineering@protocolwealthllc.com"},
]
keywords = ["llm", "gateway", "router", "openai", "ai", "compliance", "ria"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "respx>=0.22",
]

[project.scripts]
pw-router = "pw_router.__main__:main"

[project.urls]
Homepage = "https://github.com/Protocol-Wealth/pw-router"
Repository = "https://github.com/Protocol-Wealth/pw-router"
Issues = "https://github.com/Protocol-Wealth/pw-router/issues"

[tool.setuptools.packages.find]
include = ["pw_router*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
